% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfa_paged.R
\name{cfa_paged}
\alias{cfa_paged}
\title{CFA themed report output format}
\usage{
cfa_paged(
  ...,
  template = "default",
  paged_img = "default",
  css = "default",
  csl = NULL,
  toc = TRUE,
  toc_depth = 2L,
  number_sections = FALSE,
  self_contained = TRUE,
  anchor_sections = FALSE,
  mathjax = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{...}{Arguments passed to \code{bookdown::\link[bookdown]{html_document2}}
and/or \code{rmarkdown::\link[rmarkdown]{html_document_base}}.}

\item{template}{The path to the Pandoc template to convert Markdown to HTML.
The \code{'default'} will use the 'cfa_paged.html' template from this package.
Additionally, the character string \code{'appendix'} may be provided to
use the package 'cfa_appendix.html' template.}

\item{paged_img}{The images to be used as CSS variables for the cover, footer,
and other uses. Defaults are 'cfa-front-cover.svg', 'cfa-footer.svg', and
'cfa-logo.svg'. Note that if customized this should be a named character string
to take advantage of default uses. For example: \code{c(`front-cover` =
'path/to/my-front-cover.svg', `footer` = 'path/to/my-footer.svg',
`logo` = 'path/to/my-logo.svg')}.}

\item{css}{A character vector of CSS and Sass file paths. If a path
does not contain the \file{.css}, \file{.sass}, or \file{.scss} extension,
it is assumed to be a built-in CSS file. The package \code{'default'}
are \code{c('cfa-page.css', 'cfa-default.css', 'cfa-style.css')}.}

\item{csl}{The path of the Citation Style Language (CSL) file used to format
citations and references (see the
\href{https://pandoc.org/MANUAL.html#citations}{Pandoc documentation}).}

\item{toc}{Boolean to include a table of contents in the output. Default: TRUE}

\item{toc_depth}{Depth of headers to include in table of contents. Default: 2}

\item{number_sections}{Boolean whether to number section headers: if 'TRUE',
figure/table numbers will be of the form 'X.i', where 'X' is the current first-level
number, and 'i' is an incremental number (the i-th figure/table); if 'FALSE',
section figures/tables will be numbered sequentially in the document from 1, 2,
..., and you cannot cross-reference section headers in this case.}

\item{self_contained}{Produce a standalone HTML file with no external dependencies,
using data: URIs to incorporate the contents of linked scripts, stylesheets,
images, and videos. Note that even for self contained documents MathJax is
still loaded externally (this is necessary because of its size). Default: TRUE}

\item{anchor_sections}{to show section anchors when mouse hovers for all headers.
A list can also be passed with 'style' and/or 'depth' to customize the behavior.
See Anchor Sections Customization section., Default: FALSE}

\item{mathjax}{Include mathjax. Passing \code{'default'} uses an https URL from
a MathJax CDN. The 'local' option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely. Default: NULL}

\item{pandoc_args}{Additional command line options to pass to pandoc, Default: NULL}
}
\value{
An R Markdown output format.
}
\description{
See the description for \code{pagedown::\link[pagedown]{html_paged}}.
This is an alternate output format created for the Center for Assessment.
}
\references{
\url{https://pagedown.rbind.io}
}
\seealso{
\code{\link[pagedown]{html_paged}}
 \code{\link[bookdown]{html_document2}}
 \code{\link[rmarkdown]{html_document_base}}
}
\author{
Adam Van Iwaarden
}
