% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReportScripts.R
\name{createReportScripts}
\alias{createReportScripts}
\title{createReportScripts}
\usage{
createReportScripts(params, content, pagedown = TRUE, bookdown = FALSE)
}
\arguments{
\item{params}{List. A named list with parameter elements to be turned into
YAML and RMD master/parent files in the creation of a CFA report.}

\item{content}{List. A named list of RMD child documents to be included in
the report. Additional information such as file order and paths also included.}

\item{pagedown}{Logical. Should pagedown files (master/parent named by
report/appendix title) be created. Default: TRUE}

\item{bookdown}{Logical. Should bookdown files (\sQuote{_bookdown_site.yml}
and \sQuote{index.Rmd}) be created. Default: FALSE}
}
\value{
Returns one or more .Rmd and/or .yml files that contain the required
 report meta-data and child file structure/order to produce an CFA themed
 report/bookdown site.
}
\description{
Create parent .Rmd and auxilary files for rendering
 CFA themed reports.
}
\details{
Create master/parent documents appropriate for use with the CFA
 RMarkdown template. Lists provided in the \code{params} and \code{content}
 arguments are typically constructed from custom and/or universal elements.
 Standard .Rmd files are also included in the package (such as a \code{R}
 session information script for use in an appendix or within a report).
}
\examples{
\dontrun{
  createReportScripts(
     params = report.params.list
     content = report.content.file.list
  )
}
}
\seealso{
\code{\link[ymlthis]{yml}}, \code{\link[ymlthis]{yml_author}}, \code{\link[ymlthis]{yml_citations}}, \code{\link[ymlthis]{yml_params}}, \code{\link[ymlthis]{yml_bookdown_opts}}, \code{\link[ymlthis]{yml_output}}, \code{\link[ymlthis]{yml_pagedown_opts}}
 \code{\link[cfaDocs]{cfa_paged}}
}
