% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_functions.R
\name{eqnNumIncrement}
\alias{eqnNumIncrement}
\title{Increase the equation number counter}
\usage{
eqnNumIncrement(advance.counter = 1)
}
\arguments{
\item{advance.counter}{Numeric. Number to be added to the present equation
number. Default is 1, advancing equation count by 1.}
}
\value{
Next equation number
}
\description{
Bump the equation number by one.  The function relies on
  \code{options("equation_counter")} in order to keep track of the last number.
}
\examples{
\dontrun{
 if(interactive()){
   options(equation_counter=1)
   getCounter("equation")
   eqnNumIncrement()
   getCounter("equation")
 }
}
}
