% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_functions.R
\name{figNum}
\alias{figNum}
\title{Get a figure number for reference}
\usage{
figNum(advance.counter = 0)
}
\arguments{
\item{advance.counter}{Numeric.  Number to be added to the present figure
number.  Default is 0, giving the number for the last figure placed/plotted.}
}
\value{
Returns a prior, current or subsequent figure number.
}
\description{
The function relies on \code{options("fig_caption_no")}
  in order to keep track of the current number.
}
\details{
Typically used for inline R code within a .Rmd document
}
\examples{
\dontrun{
 if(interactive()){
  options(fig_caption_no=1)
  figNum()
 }
}
}
