% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_functions.R
\name{figNumIncrement}
\alias{figNumIncrement}
\title{Increase the figure number counter}
\usage{
figNumIncrement(advance.counter = 1)
}
\arguments{
\item{advance.counter}{Numeric. Number to be added to the present figure
number. Default is 1, advancing figure count by 1.}
}
\value{
Next figure number
}
\description{
Bump the figure number by one.  The function relies on
  \code{options("fig_caption_no")} in order to keep track of the last number.
}
\examples{
\dontrun{
 if(interactive()){
   options(fig_caption_no=1)
   getCounter("figure")
   figNumIncrement()
   getCounter("figure")
 }
}
}
