% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_functions.R
\name{getCounter}
\alias{getCounter}
\title{Get current counter}
\usage{
getCounter(type = c("table", "figure", "equation"))
}
\arguments{
\item{type}{Character string indicating either a \sQuote{table}, 
\sQuote{figure}, or \sQuote{equation}.}
}
\value{
Returns a (numeric) counter for the specified type.
}
\description{
Returns the current counter number associated with a
table, figure, or equation
}
\examples{
\dontrun{
  if(interactive()){
    getCounter("figure")
  }
}
}
