% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers2words.R
\name{numbers2words}
\alias{numbers2words}
\title{Convert (integer) numbers to (English) words}
\usage{
numbers2words(x = 0L)
}
\arguments{
\item{x}{Integer to be converted to (English) word(s). Default is 0 (\sQuote{zero})}
}
\value{
Returns a character string with the English equivalent of the
 provided number.
}
\description{
Returns the counter associated with a table, figure, or equation
}
\details{
Taken from the internet.
 See \href{https://stackoverflow.com/a/46652141}{this Stack Overflow conversation}
 and \href{https://gist.githubusercontent.com/hack-r/22104543e2151519c41a8f0ce042b31c/raw/01731f3176b4ee471785639533d38eda4790ab77/numbers2words.r}{Gist solution}.
}
\examples{
\dontrun{
  if(interactive()){
    numbers2words(3)
  }
}
}
