% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolvePath.R
\name{resolvePath}
\alias{resolvePath}
\title{Resolve relative paths}
\usage{
resolvePath(path = ".", relation = ".")
}
\arguments{
\item{path}{Character. File path from which to find relative path.
Defaults to the current (working) directory}

\item{relation}{Character. Relationship from path to return. 
Defaults to the path provided.}
}
\value{
Returns an absolute path - the current working dierectory by default.
}
\description{
Find a file path relative to another.
}
\details{
Useful in getting the absolute path from relative paths
}
\examples{
\dontrun{
if(interactive()){
 resolvePath(getwd(), "..")
 }
}
}
