% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_functions.R
\name{setCounters}
\alias{setCounters}
\title{Set counter numbers}
\usage{
setCounters(tbl.counter = 0, fig.counter = 0, eqn.counter = 0)
}
\arguments{
\item{tbl.counter}{Numeric. Number to set to the present table number.
Default is 0, meaning the next (first) table will be number 1.}

\item{fig.counter}{Numeric. Number to set to the present figure number.
Default is 0, meaning the next (first) figure will be number 1.}

\item{eqn.counter}{Numeric. Number to set to the present equation number.
Default is 0, meaning the next (first) equation will be number 1.}
}
\value{
NULL. Sets options
}
\description{
Sets the initial table, figure and equation counter numbers.
 Can also be used to re-set counters (e.g. for an appendix). The function
 relies on \code{options("table_number")}, \code{options("fig_caption_no")}
 and \code{options("equation_counter")} as counters.
}
\details{
DETAILS
}
\examples{
\dontrun{
  if(interactive()){
   setCounters()
  }
}
}
