% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareImage.R
\name{shareImage}
\alias{shareImage}
\title{Screenshot Your Title Slide for Share Image}
\usage{
shareImage(
  slides_rmd,
  width = 600 * (191/100),
  height = 600,
  path_image = NULL
)
}
\arguments{
\item{slides_rmd}{File name (potentially with path) of slide .Rmd}

\item{width}{Output image width (in pixels/px). Default: 600 * (191/100), must be round}

\item{height}{Output image height (in pixels/px). Default: 600, must be round}

\item{path_image}{Path to new share image. Default: NULL will create a .png
file with the same path/name as `slides_rmd` with '_share_image.png' appended.}
}
\value{
file name/path of sharable png image.
}
\description{
Takes a screenshot of your title slide for sharing on Twitter
(and other social media sites).
}
\details{
Uses \code{\link[webshot2]{rmdshot}} to re-render slides in the desired
 output ratio and creates an image of the Title page. Adapted from Garrick Aden-Buie
 blog (see references).
}
\references{
\url{https://www.garrickadenbuie.com/blog/sharing-xaringan-slides/#the-perfect-share-image-ratio}
}
\seealso{
\code{\link[rmarkdown]{yaml_front_matter}}
 \code{\link[yaml]{as.yaml}}
 \code{\link[webshot2]{rmdshot}}
}
