% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_functions.R
\name{tblCap}
\alias{tblCap}
\title{Formated caption with table number}
\usage{
tblCap(caption.text = NULL, advance.counter = 1)
}
\arguments{
\item{caption.text}{PARText string to be used for caption.  If NULL, no
caption is returned, but the table counter is advanced.AM_DESCRIPTION.
Default: NULL}

\item{advance.counter}{Numeric. Number to be added to the present table
number.  Default is 1, advancing table count by 1.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Returns a formatted markdown text string for the caption,
  which includes the table number. The function relies on
  \code{options("table_number")} in order to keep track of the current
  number.  Also increases the table number BEFORE processing the caption
  using \link{tblNumIncrement}.
}
\details{
DETAILS
}
\examples{
\dontrun{
 if(interactive()){
   options(table_number=1)
   tblCap("My table caption is great.")
   # Returns '**Table 2:**  My table caption is great.'

   getCounter("table")
   getOption("table_number")
   # Returns '2'
 }
}
}
