% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_functions.R
\name{tblNum}
\alias{tblNum}
\title{Get a table number for reference}
\usage{
tblNum(advance.counter = 0)
}
\arguments{
\item{advance.counter}{Numeric. Number to be added to the present table
number. Default is 0, giving the present table.}
}
\value{
Returns a prior, current or subsequent table number.
}
\description{
The function relies on \code{options("table_number")} in 
  order to keep track of the current number.
}
\details{
Typically used for inline R code within a .Rmd document
}
\examples{
\dontrun{
 if(interactive()){
   options(table_number=1)
   tblNum()
 }
}
}
