% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_paper_paged.R
\name{w_paper_paged}
\alias{w_paper_paged}
\title{'W' paper (working/white) output format}
\usage{
w_paper_paged(
  ...,
  appendices = list(post = list(R = list(file = "r_session_info"))),
  template = "default",
  css = "default",
  csl = "default",
  number_sections = TRUE,
  self_contained = TRUE,
  anchor_sections = FALSE,
  mathjax = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{...}{Arguments passed to \code{bookdown::\link[bookdown]{html_document2}}
and/or \code{rmarkdown::\link[rmarkdown]{html_document_base}}.}

\item{appendices}{A list of appendix source file(s) information. The default is
to add a R session information appendix that is rendered after the primary
document has been knitted (post_knit). Pre-rendered appendices (HTML fragments
that use the 'wp_appendix.html' template) and .Rmd that are knitted before
the primary document knitting are also possible.}

\item{template}{The path to the Pandoc template to convert Markdown to HTML.
The \code{'default'} will use the 'wp_paged.html' template from this package.}

\item{css}{A character vector of CSS and Sass file paths. If a path
does not contain the \file{.css}, \file{.sass}, or \file{.scss} extension,
it is assumed to be a built-in CSS file. The package \code{'default'}
are \code{c('wp-page.css', 'wp-default.css', 'wp-fonts.css')}.}

\item{csl}{The path of the Citation Style Language (CSL) file used to format
citations and references (see the
\href{https://pandoc.org/MANUAL.html#citations}{Pandoc documentation}). The
\code{'default'} is the 'american-sociological-association.csl' as used by
Bauer and Landesvatter (see description).}

\item{number_sections}{Boolean whether to number section headers: if 'TRUE',
figure/table numbers will be of the form 'X.i', where 'X' is the current
first-level number, and 'i' is an incremental number (the i-th figure/table);
if 'FALSE', section figures/tables will be numbered sequentially in the document
from 1, 2, ..., and you cannot cross-reference section headers in this case.}

\item{self_contained}{Produce a standalone HTML file with no external dependencies,
using data: URIs to incorporate the contents of linked scripts, stylesheets,
images, and videos. Note that even for self contained documents MathJax is
still loaded externally (this is necessary because of its size). Default: TRUE}

\item{anchor_sections}{to show section anchors when mouse hovers for all headers.
A list can also be passed with 'style' and/or 'depth' to customize the behavior.
See Anchor Sections Customization section., Default: FALSE}

\item{mathjax}{Include mathjax. Passing \code{'default'} uses an https URL from
a MathJax CDN. The 'local' option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely. Default: NULL}

\item{pandoc_args}{Additional command line options to pass to pandoc, Default: NULL}
}
\value{
An R Markdown output format.
}
\description{
This is a basic output format appropriate for report drafts,
working/white papers, etc. and can also incorporate appendices. No frills for
when you just want to focus on the content. This template is based on the work
done by Paul Bauer and Camille Landesvatter. See their
\href{https://github.com/paulcbauer/Writing_a_reproducable_paper_in_pagedown}{GitHub repository}
and the description of \code{pagedown::\link[pagedown]{jss_paged}} for more
general information and context.
}
\references{
\url{https://pagedown.rbind.io}
}
\seealso{
\code{\link[pagedown]{jss_paged}}
 \code{\link[bookdown]{html_document2}}
 \code{\link[rmarkdown]{html_document_base}}
}
\author{
Adam Van Iwaarden
}
